DigitalizacionPartiturasApp/requirements.txt
alembic==1.13.1
bcrypt==4.1.3
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
firebase-admin==6.5.0
Flask==3.0.2
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gcloud==0.18.3
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jws==0.1.3
Mako==1.3.2
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.26.4
oauth2client==4.1.2
opencv-python-headless==4.10.0.82
packaging==24.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.10.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.2
Pyrebase4==4.3.0
python-dateutil==2.9.0.post0
python-jwt==2.0.1
requests==2.31.0
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.1
WTForms==3.1.2
